// proof-builder/types/networks.ts
export enum SupportedNetworks {
    GNOSIS = 'gnosis',
    GOERLI = 'goerli',
    MAINNET = 'mainnet',
    ARBITRUM = 'arbitrum',
    OPTIMISM = 'optimism',
    ZKSYNC = 'zksync',
    CHIADO = 'chiado',
    BNB = 'bnb',
    POLYGON = 'polygon',
    FANTOM = 'fantom',
    AVALANCHE = 'avalanche',
    ['POLYGON-ZK'] = 'polygon-zk',
    LINEA = 'linea',
    CELO = 'celo'
}

export enum Providers {
    QUICKNODE = 'quicknode',
    ONFINALITY = 'onfinality'
}